requirements.txt
python_dotenv
eod
matplotlib
pandas
jinja2